Refactor ShareDB serving
