--- packages/neoBackend/src/shareDBMiddleware/index.js
@@ -1,4 +1,0 @@
-export { identifyAgent } from './identifyAgent';
-export { identifyOwner } from './identifyOwner';
-export { accessControl } from './accessControl';
-export { getVizInfo } from './getVizInfo';
